refactor: Carousel 네이밍 GeneralCarousel로 수정
