+++ src/components/GeneralCarousel/Dots.tsx
