+++ src/components/GeneralCarousel/GeneralCarousel.style.ts
